moving code to app-buttons
