+++ packages/geoview-core/src/core/components/app-bar/app-bar.tsx
@@ -4,7 +4,7 @@ import makeStyles from '@mui/styles/makeStyles';
 
 import { List, ListItem, Panel, IconButton } from '@/ui';
 
-import { api, PayloadBaseClass } from '@/app';
+import { api } from '@/app';
 import { EVENT_NAMES } from '@/api/events/event-types';
 
 import { MapContext } from '../../app-start';
@@ -17,8 +17,6 @@ import Geolocator from './buttons/geolocator';
 import Notifications from './buttons/notifications';
 import Version from './buttons/version';
 import ExportModal from '../export/export-modal';
-import { payloadIsANotification } from '@/api/events/payloads/notification-payload';
-import { NotificationsPopover } from '../notifications/notifications-popover';
 
 const useStyles = makeStyles((theme) => ({
   appBar: {
@@ -99,9 +97,6 @@ export function Appbar({ setActivetrap }: AppbarProps): JSX.Element {
   const [ModalIsShown, setModalIsShown] = useState(false);
   const [selectedAppBarButtonId, setSelectedAppbarButtonId] = useState<string>('');
 
-  const [notificationsCount, setNotificationsCount] = useState<number>(0);
-  const [notifPopoverAnchorEl, setNotifPopoverAnchorEl] = useState<HTMLButtonElement | null>(null);
-
   const classes = useStyles();
 
   const appBar = useRef<HTMLDivElement>(null);
@@ -121,26 +116,6 @@ export function Appbar({ setActivetrap }: AppbarProps): JSX.Element {
     setModalIsShown(false);
     // this will add back focus active trap from map and focus will be on export modal
     setActivetrap(true);
-  };
-
-  const openNotificationsPopover = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setNotifPopoverAnchorEl(event.currentTarget);
-  };
-
-  const onCloseNotificationsPopover = () => {
-    setNotifPopoverAnchorEl(null);
-  };
-
-  const onNotificationAddEvent = (payload: PayloadBaseClass) => {
-    if (payloadIsANotification(payload)) {
-      setNotificationsCount(notificationsCount + 1);
-    }
-  };
-
-  const onNotificationRemoveEvent = (payload: PayloadBaseClass) => {
-    if (payloadIsANotification(payload)) {
-      setNotificationsCount(notificationsCount - 1);
-    }
   };
 
   const addButtonPanel = useCallback(
@@ -202,20 +177,12 @@ export function Appbar({ setActivetrap }: AppbarProps): JSX.Element {
       `${mapId}/${selectedAppBarButtonId}`
     );
 
-    // listen to notifications add
-    api.event.on(EVENT_NAMES.NOTIFICATIONS.NOTIFICATION_ADD, onNotificationAddEvent, mapId);
-
-    // listen to notifications add
-    api.event.on(EVENT_NAMES.NOTIFICATIONS.NOTIFICATION_REMOVE, onNotificationRemoveEvent, mapId);
-
     return () => {
       api.event.off(EVENT_NAMES.APPBAR.EVENT_APPBAR_PANEL_CREATE, mapId);
       api.event.off(EVENT_NAMES.APPBAR.EVENT_APPBAR_PANEL_REMOVE, mapId);
       api.event.off(EVENT_NAMES.PANEL.EVENT_PANEL_CLOSE, mapId);
-      api.event.off(EVENT_NAMES.NOTIFICATIONS.NOTIFICATION_REMOVE, mapId, onNotificationRemoveEvent);
-      api.event.off(EVENT_NAMES.NOTIFICATIONS.NOTIFICATION_ADD, mapId, onNotificationAddEvent);
     };
-  }, [addButtonPanel, mapId, removeButtonPanel, notificationsCount, selectedAppBarButtonId]);
+  }, [addButtonPanel, mapId, removeButtonPanel, selectedAppBarButtonId]);
 
   return (
     <div className={classes.appBar} ref={appBar}>
@@ -280,11 +247,7 @@ export function Appbar({ setActivetrap }: AppbarProps): JSX.Element {
         <div className={classes.versionButtonDiv}>
           <List className={classes.appBarList}>
             <ListItem>
-              <Notifications
-                notificationsCount={notificationsCount}
-                className={`${classes.appBarButton} ${ModalIsShown ? 'active' : ''}`}
-                openPopover={openNotificationsPopover}
-              />
+              <Notifications />
             </ListItem>
             <ListItem>
               <Version />
@@ -309,7 +272,6 @@ export function Appbar({ setActivetrap }: AppbarProps): JSX.Element {
         );
       })}
       <ExportModal isShown={ModalIsShown} closeModal={closeModal} />
-      <NotificationsPopover anchorEl={notifPopoverAnchorEl} handleClose={onCloseNotificationsPopover} />
     </div>
   );
 }
